Organise countries + add pinning feature
Sorts countries alphabetically, pins UK, US, AUS and CA, and enables any countries that have state data in the seed. If someone has a lot of countries enabled, the list can get pretty long. This commit adds code that enables the user to pin favourite countries to the top of the list.