Update Estonia to use meaningful codes
The code field does not represent an ISO standard, the intention is for it to be used in slugs and other codes to be read by humans. If we want ISO codes, then we should add a dedicated field for them. Meaningful codes sourced from: http://www.geopostcodes.com/Estonia

Refs https://github.com/rainlab/location-plugin/pull/24